Rename Toolbox
Test Regression tool, and minor edits to metadata. Generally, I am
tempted to remove the regression tool, but it still works and its use of
reflection is still relatively clever. Not great for actual use, but
does provide a nice sample on how to get useful variables from the
models generated.